Provide an option for fuzzy matching: ido-at-point-fuzzy
For example, this would suggest "ido-at-point-complete" for the query
"iapc"." This is similar to the functionality that flx-ido provides
where you can access a function quickly via its abbreviation.

The use of the fuzzy matching depends on `ido-at-point-fuzzy`; the
functionality of the package is identical unless the user changes
`ido-at-point-fuzzy`.